Adding trimIndent implementation
The "trimIndent" function's implementation is now complete. However, it
seems that its usage will not be necessary, as template strings with
linebreaks do not incur in actual linebreaks being shown in the page, as
was expected when the function was thought of. Still, to preserve the
implementation, for now, the file is kept. Should it prove unnecessary
within a reasonable number of commits, it should be removed.